% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_report}
\alias{download_report}
\title{Download win-devel report}
\usage{
download_report(url, path.dump = file.path(getwd(), "check_win_devel"))
}
\arguments{
\item{url}{character; url like "win-builder.r-project.org/..."}

\item{path.dump}{character; local path to store the download, the DEFAULT is
the folder \code{check_win_devel} at the current working directory.}
}
\description{
Download win-devel report
}
\examples{
\dontrun{
download_report("https://win-builder.r-project.org/8esWmBov13K0")
# or
download_report("8esWmBov13K0")
}

}
