% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_folder_name}
\alias{get_folder_name}
\title{Get the parent folder of a url folder}
\usage{
get_folder_name(path)
}
\arguments{
\item{path}{character; a url path, must be a folder, i.e., to end with "/".}
}
\value{
character.
}
\description{
Get the parent folder of a url folder
}
