% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{init}
\alias{init}
\title{Initializes a folder to save the downloads and adds it to the \code{.Rbuildignore}}
\usage{
init()
}
\description{
Initializes a folder to save the downloads and adds it to the \code{.Rbuildignore}
}
