% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{is_folder}
\alias{is_folder}
\title{Verify if elements of a vector of URLs redirects to a folder or a file}
\usage{
is_folder(vec_of_urls)
}
\arguments{
\item{vec_of_urls}{character vector; URLs to verify if they're folder or
not.}
}
\value{
boolean vector. Each element is
\itemize{
\item \code{TRUE} if it's a folder;
\item \code{FALSE} if it's a file.
}
}
\description{
Verify if elements of a vector of URLs redirects to a folder or a file
}
