% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{validate_url}
\alias{validate_url}
\title{Validate parameters of the function \code{download_report()}}
\usage{
validate_url(url)
}
\arguments{
\item{url}{character; url.}
}
\value{
character; url standardized.
}
\description{
Validate parameters of the function \code{download_report()}
}
